Final answer

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
 */
 size_t print_list(const list_t *h)
 {
-	int count_node = 0;
+	size_t count_node = 0;
 
 	while (h)
 	{
